+++ dodo.py
@@ -122,11 +122,8 @@ def task_setup():
     for pkg, pkg_setup in P.PY_SETUP.items():
         # TODO: refactor
         ext_deps = [
-            pkg_setup.parent
-            / "src"
-            / pkg.replace("-", "_")
-            / "labextensions"
-            / P.JS_PKG_DATA[ext]["name"]
+            P.JS_PKG_JSON[ext].parent
+            / P.JS_PKG_DATA[ext]["jupyterlab"]["outputDir"]
             / "package.json"
             for ext, mod in P.JS_LABEXT_PY_HOST.items()
             if mod == pkg_setup.parent.name
@@ -377,9 +374,7 @@ if not P.TESTING_IN_CI:
             if "jupyterlab" not in pkg_data:
                 continue
 
-            host = P.JS_LABEXT_PY_HOST[pkg]
-            host_mod = host.replace("-", "_")
-            host_ext = P.PY_PACKAGES / host / "src" / host_mod / "labextensions"
+            out_dir = pkg_data["jupyterlab"]["outputDir"]
 
             yield _ok(
                 dict(
@@ -392,19 +387,15 @@ if not P.TESTING_IN_CI:
                         )
                     ],
                     file_dep=targets,
-                    targets=[host_ext / f"""{pkg_data["name"]}/package.json"""],
+                    targets=[P.JS_PKG_JSON[pkg] / f"{out_dir}/package.json"],
                 ),
                 P.OK_EXT_BUILD[pkg],
             )
 
         for py_pkg, py_setup in P.PY_SETUP.items():
-            py_mod = py_setup.parent.name.replace("-", "_")
             ext_deps = [
-                py_setup.parent
-                / "src"
-                / py_mod
-                / "labextensions"
-                / P.JS_PKG_DATA[ext]["name"]
+                P.JS_PKG_JSON[ext].parent
+                / P.JS_PKG_DATA[ext]["jupyterlab"]["outputDir"]
                 / "package.json"
                 for ext, mod in P.JS_LABEXT_PY_HOST.items()
                 if mod == py_setup.parent.name