+++ py_packages/ipydrawio/src/ipydrawio/__init__.py
@@ -28,7 +28,7 @@ def _jupyter_labextension_paths():
             src=f"{pkg.parent.relative_to(here).as_posix()}",
             dest=f"{pkg.parent.parent.name}/{pkg.parent.name}",
         )
-        for pkg in (here / "labextensions").glob("*/*/package.json")
+        for pkg in here.glob("ext/*/package.json")
     ]